ukk 2

+++ app/Http/Controllers/RoomController.php
@@ -7,6 +7,7 @@ use App\Models\RoomImage;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Http\Request;
 
+
 class RoomController extends Controller
 {
     public function index()
@@ -24,11 +25,12 @@ class RoomController extends Controller
         return view('admin.dashboard', compact('rooms'));
     }
 
-    public function dashboard()
+    public function RecepsionistIndex()
     {
-        $rooms = Room::all();
-        return view('admin.dashboard', compact('rooms'));
+        $rooms = Room::latest()->paginate(5)->withQueryString();
+        return view('receptionist.kamar', compact('rooms'))->with('i', (request()->input('page', 1) - 1) * 5);
     }
+
     public function create()
     {
         return view('admin.room.create');
@@ -45,7 +47,7 @@ class RoomController extends Controller
         ]);
 
         $room = Room::create([
-            'rooms_id' => \Illuminate\Support\Str::uuid(), // Pastikan UUID dibuat
+            'rooms_id' => \Illuminate\Support\Str::uuid(),
             'room_type' => $request->room_type,
             'price' => $request->price,
             'total_room' => $request->total_room,
@@ -54,13 +56,12 @@ class RoomController extends Controller
             'image' => $request->file('image') ? $request->file('image')->store('room_images', 'public') : null,
         ]);
 
-        // Simpan gambar jika ada yang diunggah
         if ($request->hasFile('images')) {
             foreach ($request->file('images') as $image) {
                 $imagePath = $image->store('rooms', 'public');
 
                 RoomImage::create([
-                    'rooms_id' => $room->rooms_id, // Pastikan sesuai dengan primary key Room
+                    'rooms_id' => $room->rooms_id,
                     'image_path' => $imagePath,
                 ]);
             }
@@ -96,13 +97,19 @@ class RoomController extends Controller
 
         return redirect()->route('admin.dashboard')->with('success', 'Room updated successfully!');
     }
-    public function destroy($id)
-    {
-        $room = Room::findOrFail($id);
-        $room->delete();
+    public function destroy($room_id)
+{
+    $room = Room::findOrFail($room_id);
 
-        return redirect()->route('admin.room.destroy')->with('success', 'Room deleted successfully!');
+    if ($room->reservations()->exists()) {  
+        return redirect()->back()->with('error', 'Room tidak bisa dihapus karena masih ada pesanan.');
     }
+
+    $room->delete();
+
+    return redirect()->route('admin.room.index')->with('success', 'Room berhasil dihapus!');
+}
+
     public function show($id)
     {
         $room = Room::with(['images'])->findOrFail($id);
